Added moving platforms, a stationary ground(for testing purposes), additional control keys, and two level styles
The moving platforms are currently in machine gun mode. I'm not sure how to separate the levels yet, so the shotgun level is currently just commented out. I added a sprint key, along with the option to use a and d to control left and right movement, but left and right arrow keys still work as well. The ground is a temporary addition to allow us extra time to test what we need to without the fear of falling.